src/tdd_practices/arrays/sum_test.go
package main

import (
	"reflect"
	"testing"
)

func TestSum(t *testing.T) {
	t.Run("collection of 5 numbers", func(t *testing.T) {
		// [N] fixed capacity
		numbers := []int{1, 2, 3, 4, 5}

		got := Sum(numbers)
		want := 15

		if got != want {
			// %v print default format => works well for arrays
			t.Errorf("got %d want %d given, %v", got, want, numbers)
		}
	})

	// use slice!
	// slice can be resized
	t.Run("collection of any size", func(t *testing.T) {
		// use slice []
		numbers := []int{1, 2, 3}

		got := Sum(numbers)
		want := 6

		if got != want {
			// %v print default format => works well for arrays
			t.Errorf("got %d want %d given, %v", got, want, numbers)
		}
	})
}

func TestSumAll(t *testing.T) {
	got := SumAll([]int{1, 2}, []int{0, 9})
	want := []int{3, 9}

	// Go does not allow to use equality operators with slice
	// use reflect DeepEqual to see if any 2 variables are the same (NOT type safe!)
	if !reflect.DeepEqual(got, want) {
		t.Errorf("got %v want %v", got, want)
	}
}

func TestSumAllTails(t *testing.T) {
	t.Run("make the sums of some slices", func(t *testing.T) {
		got := SumAllTails([]int{1, 2}, []int{0, 9})
		want := []int{2, 9}

		if !reflect.DeepEqual(got, want) {
			t.Errorf("got %v want %v", got, want)
		}
	})

	t.Run("safely sum empty slices", func(t *testing.T) {
		got := SumAllTails([]int{}, []int{3, 4, 5})
		want := []int{0, 9}

		if !reflect.DeepEqual(got, want) {
			t.Errorf("got %v want %v", got, want)
		}
	})
}

// go test -cover for coverage
